backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blockbuster==1.5.25
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorama==0.4.6
cryptography==44.0.3
dataclasses-json==0.6.7
distro==1.9.0
forbiddenfruit==0.1.4
frozenlist==1.8.0
genai-prices==0.0.44
googleapis-common-protos==1.72.0
griffe==1.15.0
groq==0.36.0
grpcio==1.76.0
grpcio-tools==1.75.1
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain-classic==1.0.0
langchain-core==1.0.5
langchain-tavily==0.2.13
langchain-text-splitters==1.0.0
langgraph-api==0.4.48
langgraph-checkpoint==2.1.2
langgraph-checkpoint-postgres>=2.0.0
langgraph-prebuilt==1.0.4
langgraph-runtime-inmem==0.14.1
langgraph-sdk==0.2.9
logfire-api==4.15.1
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
propcache==0.4.1
protobuf==6.33.1
pycparser==2.23
pydantic==2.12.4
pydantic-ai-slim==1.22.0
pydantic-graph==1.22.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.0.0
PyJWT==2.10.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.44
psycopg-binary==3.3.2
sse-starlette==2.1.3
starlette==0.49.3
structlog==25.5.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
truststore==0.10.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
watchfiles==1.1.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
